Speech-to-Text-cf-http-endpoint/requirements.txt
#requirements.txt
functions-framework
firebase-admin
google-cloud-speech